Split errs into module with local and foregin impls
